Added possibility in CachedBlock to also get empty blocks. Slightly optimized the block allocation code in the case that the maximum allocation is equal to the minimum allocation. The allocation now cuts the requested number of blocks to the allocation groups boundaries. Implemented growing the data stream in the indirect range - tested. Fixed some bugs in Inode::ShrinkStream() for the indirect range. Updated WSTAT_MODE that it does at least something - but the whole function is not to nice in checking permissions and stuff.

+++ BeFS/source/Volume.h
@@ -66,7 +66,7 @@ class Volume {
 
 		status_t			AllocateForInode(Transaction *transaction,const Inode *parent,mode_t type,block_run &run);
 		status_t			AllocateForInode(Transaction *transaction,const block_run *parent,mode_t type,block_run &run);
-		status_t			Allocate(Transaction *transaction,const Inode *inode,uint16 numBlocks,block_run &run,uint16 minimum = 1);
+		status_t			Allocate(Transaction *transaction,const Inode *inode,off_t numBlocks,block_run &run,uint16 minimum = 1);
 		status_t			Free(Transaction *transaction,block_run &run);
 
 #ifdef DEBUG
@@ -95,7 +95,7 @@ Volume::AllocateForInode(Transaction *transaction, const block_run *parent, mode
 
 
 inline status_t 
-Volume::Allocate(Transaction *transaction, const Inode *inode, uint16 numBlocks, block_run &run, uint16 minimum)
+Volume::Allocate(Transaction *transaction, const Inode *inode, off_t numBlocks, block_run &run, uint16 minimum)
 {
 	return fBlockAllocator.Allocate(transaction,inode,numBlocks,run,minimum);
 }
